requirements.txt
appdirs==1.4.4
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
fastobo==0.14.1
fonttools==4.60.1
kiwisolver==1.4.9
lark==1.3.1
llvmlite==0.45.1
matplotlib==3.10.7
numba==0.62.1
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
pyteomics==4.7.5
python-dateutil==2.9.0.post0
six==1.17.0
spectrum-utils==0.4.2
sqlparse==0.5.3
